strlen

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: david <david@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/23 22:05:50 by david         #+#    #+#                 */
-/*   Updated: 2022/03/16 13:17:43 by dyeboa        ########   odam.nl         */
+/*   Updated: 2022/03/16 13:49:31 by dyeboa        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s && s[i])
